+++ src/components/board/view/aiComment/AICommentItem.module.css
@@ -1,0 +1,72 @@
+.aiCommentContainer {
+    padding: 20px;
+    background-color: #f8f9fa;
+    border-radius: 8px;
+}
+
+.interactionBlock {
+    margin-bottom: 20px;
+}
+
+.aiAnswer {
+    padding: 16px;
+    background-color: #f1f1f1;
+    border-radius: 6px;
+    margin-bottom: 12px;
+}
+
+.followUpQuestion {
+    padding: 16px;
+    background-color: white;
+    border-radius: 6px;
+    margin-left: 20px;
+    border-left: 4px solid #666;
+}
+
+.questionForm {
+    margin-top: 24px;
+    padding: 16px;
+    background-color: white;
+    border-radius: 6px;
+    box-shadow: 0 2px 4px rgba(0, 0, 0, 0.1);
+}
+
+.inputGroup {
+    display: flex;
+    gap: 12px;
+    margin-top: 12px;
+}
+
+.input {
+    flex: 1;
+    padding: 8px 12px;
+    border: 1px solid #ddd;
+    border-radius: 4px;
+    font-size: 14px;
+}
+
+.button {
+    padding: 8px 16px;
+    background-color: #666;
+    color: white;
+    border: none;
+    border-radius: 4px;
+    cursor: pointer;
+    transition: background-color 0.2s;
+}
+
+.button:hover {
+    background-color: #444;
+}
+
+.h3 {
+    margin: 0 0 12px 0;
+    font-size: 16px;
+    color: #333;
+}
+
+.p {
+    margin: 0;
+    line-height: 1.5;
+    color: #555;
+}
